+++ src/main/java/com/rakovets/course/java/core/practice/oop_principles/battle_ground/enemies/Werewolf.java
@@ -43,8 +43,4 @@ public class Werewolf extends Enemy {
             }
         }
     }
-
-    public boolean isAlive() {
-        return this.getHealth() > 0;
-    }
 }
